+++ app/views/musica/index.php
@@ -51,15 +51,17 @@
                                     class="btn btn-primary">
                                     <i class="fas fa-eye me-1"></i>Ver
                                 </a>
-                                <a href="<?php echo $url_alias; ?>/musica/update/<?php echo $musica['id_musica']; ?>"
-                                    class="btn btn-warning">
-                                    <i class="fas fa-edit me-1"></i>Editar
-                                </a>
-                                <a href="<?php echo $url_alias; ?>/musica/delete/<?php echo $musica['id_musica']; ?>"
-                                    class="btn btn-danger"
-                                    onclick="return confirm('Tem certeza que deseja eliminar esta música?');">
-                                    <i class="fas fa-trash-alt me-1"></i>Eliminar
-                                </a>
+                                <?php if ($isAdmin): ?>
+                                    <a href="<?php echo $url_alias; ?>/musica/update/<?php echo $musica['id_musica']; ?>"
+                                        class="btn btn-warning">
+                                        <i class="fas fa-edit me-1"></i>Editar
+                                    </a>
+                                    <a href="<?php echo $url_alias; ?>/musica/delete/<?php echo $musica['id_musica']; ?>"
+                                        class="btn btn-danger"
+                                        onclick="return confirm('Tem certeza que deseja eliminar esta música?');">
+                                        <i class="fas fa-trash-alt me-1"></i>Eliminar
+                                    </a>
+                                <?php endif; ?>
                             </div>
                         </div>
                     </div>
